src/requirements.txt
numpy==1.22.4
pandas==1.3.5
scikit-learn==1.2.2
scipy==1.10.1
torch==1.10.0
torchtext==0.11.0
transformers==4.15.0
wordninja==2.0.0
tweet-preprocessor==0.6.0
gensim==3.8.3
